# app/session6-testing/api.yml
openapi: 3.0.3
info:
  title: Mini EASM API - Session 6 (Testing)
  description: |
    # External Attack Surface Management API

    API hoàn chỉnh cho hệ thống EASM với các tính năng:
    - ✅ Asset Management (CRUD operations)
    - ✅ Scanning Operations (DNS, WHOIS, Subdomain)
    - ✅ Scan Job Management
    - ✅ Result Retrieval & Filtering
    - ✅ Pagination Support
    - 🧪 Full Unit Test Coverage (26 tests + 6 benchmarks)

    ## Sessions Progression:
    - Session 2: Basic API (CRUD)
    - Session 3: Database Integration (PostgreSQL)
    - Session 4: Advanced API (Filtering, Pagination, Validation)
    - Session 5: EASM Scanning (DNS, WHOIS, Subdomain Enumeration)
    - Session 6: Testing & Quality Assurance ← Current

    ## Scan Types:

    ### 🟢 Passive Scans (Safe - No Permission Required)
    - `dns` - DNS record queries (A, AAAA, MX, NS, TXT, CNAME)
    - `whois` - WHOIS registration information
    - `subdomain` - Subdomain enumeration via DNS bruteforce
    - `cert_trans` - Certificate Transparency logs
    - `asn` - ASN lookup
    - `all` - Run all passive scans

    ### 🔴 Active Scans (⚠️ Requires Permission)
    - `port` - Port scanning (requires authorization)
    - `ssl` - SSL/TLS certificate probing (requires authorization)

    **⚠️ LEGAL WARNING:** Active scans may be illegal without explicit permission from target owner.
    Only scan systems you own or have written authorization to test.

  version: 6.0.0
  contact:
    name: CMC Intern Program
...
servers:
...
paths:
...
  /assets:
    post:
      summary: Create new asset
      description: |
        Tạo mới một asset (domain, IP, hoặc service).

        **Validation Rules:**
        - Name: required, max 255 characters, no null bytes
        - Type: must be "domain", "ip", or "service"
        - Domain: valid format (e.g., example.com), no leading/trailing dots
        - IP: valid IPv4 or IPv6 format
        - Service: valid URL format

      operationId: createAsset
      tags:
        - Assets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: example.com
                  description: Tên của asset (domain, IP, hoặc URL)
                type:
                  type: string
                  enum: [domain, ip, service]
                  example: domain
                  description: Loại asset
            examples:
              domain:
                summary: Domain asset
                value:
                  name: example.com
                  type: domain
              ip:
                summary: IP address asset
                value:
                  name: 192.168.1.1
                  type: ip
              service:
                summary: Service/URL asset
                value:
                  name: https://api.example.com
                  type: service
      responses:
        "201":
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input (validation error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                empty_name:
                  summary: Empty name
                  value:
                    error: "name is required"
                invalid_type:
                  summary: Invalid type
                  value:
                    error: "invalid asset type: must be domain, ip, or service"
                invalid_domain:
                  summary: Invalid domain format
                  value:
                    error: "invalid domain: invalid domain format (e.g., example.com)"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List assets with filtering and pagination
      description: |
        Lấy danh sách assets với các tính năng:
        - 📋 Pagination (page, page_size)
        - 🔍 Filtering (type, status)
        - 🔎 Search (partial match on name)
        - 🔄 Sorting (sort_by, sort_order)

        **Query Parameters:**
        - `page`: Page number (default: 1)
        - `page_size`: Items per page (default: 20, max: 100)
        - `type`: Filter by asset type (domain, ip, service)
        - `status`: Filter by status (active, inactive)
        - `search`: Search in asset name (case-insensitive partial match)
        - `sort_by`: Sort field (name, type, status, created_at, updated_at)
        - `sort_order`: Sort order (asc, desc)

      operationId: listAssets
      tags:
        - Assets
      parameters:
        - name: page
          in: query
          description: Page number (starts from 1)
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: page_size
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: type
          in: query
          description: Filter by asset type
          schema:
            type: string
            enum: [domain, ip, service]
            example: domain
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, inactive]
            example: active
        - name: search
          in: query
          description: Search term (partial match on name)
          schema:
            type: string
            example: example
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [name, type, status, created_at, updated_at]
            default: created_at
            example: name
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: asc
      responses:
        "200":
          description: Paginated list of assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
                  total:
                    type: integer
                    example: 150
                    description: Total number of items matching filters
                  page:
                    type: integer
                    example: 1
                  page_size:
                    type: integer
                    example: 20
                  total_pages:
                    type: integer
                    example: 8
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /assets/{id}:
    get:
      summary: Get single asset by ID
      description: Lấy thông tin chi tiết của một asset
      operationId: getAsset
      tags:
        - Assets
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Asset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "asset not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update asset
      description: |
        Cập nhật thông tin của một asset.
        Tất cả fields đều optional - chỉ cần gửi fields muốn update.
      operationId: updateAsset
      tags:
        - Assets
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: new-example.com
                type:
                  type: string
                  enum: [domain, ip, service]
                  example: domain
                status:
                  type: string
                  enum: [active, inactive]
                  example: inactive
            examples:
              update_status:
                summary: Update status only
                value:
                  status: inactive
              update_all:
                summary: Update all fields
                value:
                  name: updated-example.com
                  type: domain
                  status: active
      responses:
        "200":
          description: Asset updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete asset
      description: Xóa một asset (soft delete hoặc hard delete tùy implementation)
      operationId: deleteAsset
      tags:
        - Assets
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Asset deleted successfully (no content)
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================
  # SCANNING OPERATIONS
  # ============================================
...
components:
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      description: Asset UUID
      schema:
        type: string
        format: uuid
        example: "123e4567-e89b-12d3-a456-426614174000"
...
  schemas:
...
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: example.com
          description: Asset name (domain, IP, or URL)
        type:
          type: string
          enum: [domain, ip, service]
          example: domain
        status:
          type: string
          enum: [active, inactive]
          example: active
        created_at:
          type: string
          format: date-time
          example: "2026-03-09T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-03-09T10:30:00Z"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "invalid request body"
          description: Error message describing what went wrong